List of types, structs and functions

Types

CountTimeSeries.CountModel — Type

CountModel

Abstract type which summarizes all models in the CountTimeSeries.jl package.

CountTimeSeries.INGARCH — Type

INGARCH

Abstract type which summarizes all models in the INGARCH framework. Subtype of CountModel.

CountTimeSeries.INARMA — Type

INARMA

Abstract type which summarizes all models in the INARMA framework. Subtype of CountModel.

Structs

CountTimeSeries.INGARCHModel — Type

INGARCHModel

Struct to specify an INGARCH model. It contains information on

Distribution

Link function

Past observations included in the conditional mean definition

Past mean included in the conditional mean definition

Regressor matrix X

external: Indicator which regressors enter the system externally

zi: Indicator. Is zero inflation modelled?

CountTimeSeries.INARCHModel — Type

INARCHModel

Struct to specify an INARCH model. It contains information on

Distribution

Link function

Past observations included in the conditional mean definition

Regressor matrix X

external: Indicator which regressors enter the system externally

zi: Indicator. Is zero inflation modelled?

CountTimeSeries.IIDModel — Type

IIDModel

Struct to specify an IID model. It contains information on

Distribution

Link function

Regressor matrix X

external: Indicator which regressors enter the system externally

zi: Indicator. Is zero inflation modelled?

CountTimeSeries.INARMAModel — Type

INARMAModel

Struct to specify an INARMA model. It contains information on

Distributions

Link functions

Past observations included

Past mean included

Regressor matrix X

external: Indicator which regressors enter the system externally

zi: Indicator. Is zero inflation modelled?

CountTimeSeries.INARModel — Type

INARModel

Struct to specify an INAR model. It contains information on

Distributions

Link functions

Past observations included

Regressor matrix X

external: Indicator which regressors enter the system externally

zi: Indicator. Is zero inflation modelled?

CountTimeSeries.INMAModel — Type

INMAModel

Struct to specify an INMA model. It contains information on

Distributions

Link functions

Past mean included

Regressor matrix X

external: Indicator which regressors enter the system externally

zi: Indicator. Is zero inflation modelled?

CountTimeSeries.parameter — Type

parameter(β0, α, β, η, ϕ, ω)

Structure for Count Data models parameters.

β0: Intercept parameter

α: Parameters for autoregression

β: Parameters for MA part/past means

η: Regressor parameters

ϕ: Overdispersion parameter (Negative Binomial)

ω: Zero inflation probability

For details, see For details, see Documentation

Functions

Missing docstring.

Missing docstring for AIC(results, dropfirst). Check Documenter's build log for details.

Missing docstring.

Missing docstring for BIC(results, dropfirst). Check Documenter's build log for details.

Missing docstring.

Missing docstring for HQIC(results, dropfirst). Check Documenter's build log for details.

Missing docstring.

Missing docstring for fit(y, model, MLEControl, printResults = true, initiate = "first"). Check Documenter's build log for details.

Missing docstring.

Missing docstring for ll(y, model, θ; initiate = "first"). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MLESettings(y, model, init, optimizer, ci). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Model(model, distr, link, pastObs, pastMean, X, external, zi). Check Documenter's build log for details.

Missing docstring.

Missing docstring for par2θ(θ, model). Check Documenter's build log for details.

Missing docstring.

Missing docstring for pit(results, nbins, level). Check Documenter's build log for details.

Missing docstring.

Missing docstring for predict(results, h, nChain, Xnew). Check Documenter's build log for details.

Missing docstring.

Missing docstring for QPois(results). Check Documenter's build log for details.

Missing docstring.

Missing docstring for simulate(T, model, θ; burnin, pinfirst). Check Documenter's build log for details.

Missing docstring.

Missing docstring for θ2par(θ, model). Check Documenter's build log for details.

Base.:∘ — Function

f ∘ g

Compose functions: i.e. (f ∘ g)(args...) means f(g(args...)). The ∘ symbol can be entered in the Julia REPL (and most editors, appropriately configured) by typing \circ<tab>.

Function composition also works in prefix form: ∘(f, g) is the same as f ∘ g. The prefix form supports composition of multiple functions: ∘(f, g, h) = f ∘ g ∘ h and splatting ∘(fs...) for composing an iterable collection of functions.

Julia 1.4

Multiple function composition requires at least Julia 1.4.

Julia 1.5

Composition of one function ∘(f) requires at least Julia 1.5.

Examples

julia> map(uppercase∘first, ["apple", "banana", "carrot"])
3-element Array{Char,1}:
 'A': ASCII/Unicode U+0041 (category Lu: Letter, uppercase)
 'B': ASCII/Unicode U+0042 (category Lu: Letter, uppercase)
 'C': ASCII/Unicode U+0043 (category Lu: Letter, uppercase)

julia> fs = [
           x -> 2x
           x -> x/2
           x -> x-1
           x -> x+1
       ];

julia> ∘(fs...)(3)
3.0